+++ src/components/Avatar/Avatar.tsx
@@ -7,28 +7,25 @@ import { elementColorsNames, getElementColor } from "../../utils/colors-vars-map
 import { AvatarSizes, AvatarTypes } from "./AvatarConstants";
 import { AvatarBadge, AvatarBadgeProps } from "./AvatarBadge";
 import { AvatarContent } from "./AvatarContent";
-import Tooltip from "../Tooltip/Tooltip";
+import Tooltip, { TooltipProps } from "../Tooltip/Tooltip";
 import ClickableWrapper from "../Clickable/ClickableWrapper";
 import Dialog from "../Dialog/Dialog";
-import { iconSubComponentProps } from "../Icon/Icon";
-
-import "./Avatar.scss";
 import VibeComponentProps from "src/types/VibeComponentProps";
+import { IconSubComponentProps } from "../Icon/Icon";
+import "./Avatar.scss";
 
 const AVATAR_CSS_BASE_CLASS = "monday-style-avatar";
 const bemHelper = BEMClass(AVATAR_CSS_BASE_CLASS);
 
 type BackgroundColors = typeof elementColorsNames[keyof typeof elementColorsNames];
 export interface AvatarProps extends VibeComponentProps {
-  id?: string;
   src?: string;
   text?: string;
-  tooltipProps?: any;
+  tooltipProps?: TooltipProps;
   ariaLabel?: string;
   withoutTooltip?: boolean;
-  icon?: string | React.FunctionComponent<iconSubComponentProps> | null;
+  icon?: string | React.FunctionComponent<IconSubComponentProps> | null;
   type?: AvatarTypes;
-  className?: string;
   textClassName?: string;
   backgroundColor?: BackgroundColors;
   customBackgroundColor?: string;
@@ -207,9 +204,7 @@ Object.assign(Avatar, {
 });
 
 Avatar.defaultProps = {
-  id: undefined,
   src: undefined,
-  className: "",
   textClassName: "",
   icon: undefined,
   text: undefined,