+++ webpack/published-ts-components.js
@@ -1,6 +1,7 @@
 const publishedTSComponents = {
+  // Don't remove next line
+  // plop_marker:published-components
   Button: "components/Button/Button",
-  AlertBannerButton: "components/AlertBanner/AlertBannerButton/AlertBannerButton",
   ToastButton: "components/Toast/ToastButton/ToastButton",
   Icon: "components/Icon/Icon",
   IconButton: "components/IconButton/IconButton",
@@ -12,13 +13,31 @@ const publishedTSComponents = {
   HiddenText: "components/HiddenText/HiddenText",
   Label: "components/Label/Label",
   RadioButton: "components/RadioButton/RadioButton",
+  DialogContentContainer: "components/DialogContentContainer/DialogContentContainer",
   Banner: "components/Banner/Banner",
+  ButtonGroup: "components/ButtonGroup/ButtonGroup",
   Clickable: "components/Clickable/Clickable",
+  AlertBanner: "components/AlertBanner/AlertBanner",
+  AlertBannerButton: "components/AlertBanner/AlertBannerButton/AlertBannerButton",
+  AlertBannerText: "components/AlertBanner/AlertBannerText/AlertBannerText",
   AlertBannerLink: "components/AlertBanner/AlertBannerLink/AlertBannerLink",
   ToastLink: "components/Toast/ToastLink/ToastLink",
+  Tooltip: "components/Tooltip/Tooltip",
   CustomSvgIcon: "components/Icon/CustomSvgIcon/CustomSvgIcon",
   Avatar: "components/Avatar/Avatar",
   Chips: "components/Chips/Chips",
+  BreadcrumbsBar: "components/BreadcrumbsBar/BreadcrumbsBar",
+  BreadcrumbItem: "components/BreadcrumbsBar/BreadcrumbItem/BreadcrumbItem",
+  Heading: "components/Heading/Heading",
+  TextWithHighlight: "components/TextWithHighlight/TextWithHighlight",
+  List: "components/List/List",
+  ListItem: "components/ListItem/ListItem",
+  ListTitle: "components/ListTitle/ListTitle",
+  ListItemIcon: "components/ListItemIcon/ListItemIcon",
+  Loader: "components/Loader/Loader",
+  AttentionBox: "components/AttentionBox/AttentionBox",
+  SearchComponent: "components/Search/Search", // TODO: remove when bumping to version 1.0.0
+  Search: "components/Search/Search",
   useKeyEvent: "hooks/useKeyEvent/index.ts",
   useEventListener: "hooks/useEventListener/index.ts",
   useDebounceEvent: "hooks/useDebounceEvent/index.ts",
@@ -35,6 +54,8 @@ const publishedTSComponents = {
   useMediaQuery: "hooks/useMediaQuery",
   useVibeMediaQuery: "hooks/useVibeMediaQuery",
   useActiveDescendantListFocus: "hooks/useActiveDescendantListFocus"
+  // Don't remove next line
+  // plop_marker:published-hooks
 };
 
 // eslint-disable-next-line no-undef